arranged files for .h and .cpp to make coding more clearer
