--- Command.cpp
@@ -1,4 +1,0 @@
-class Command {
-	public:
-  virtual int execute(string str) = 0;
-};
