+++ Commands.cpp
@@ -1,0 +1,7 @@
+//file to write all Commands implementation
+#include Command.h
+#include ConnectCommand.h
+#include DefineVarCommand.h
+#include LoopCommand.h
+#include OpenServerCommand.h
+#include SleepCommand.h
