--- ConnectCommand.cpp
@@ -1,6 +1,0 @@
-class ConnectCommand : Command {
-  int execute(string str) {
-
-  }
-  ~ConnectCommand();
-};
