+++ ConnectCommand.h
@@ -1,0 +1,11 @@
+#ifndef FLIGHTSIMULATOR__ConnectCOMMAND_H_
+#define FLIGHTSIMULATOR__ConnectCOMMAND_H_
+#include Command.h
+class ConnectCommand : Command {
+	public:
+  int execute(string str) {
+
+  }
+  virtual ~ConnectCommand();
+};
+#endif
