--- DefineVarCommand.cpp
@@ -1,6 +1,0 @@
-class DefineVarCommand : Command {
-int execute(string str) {
-
-}
-~DefineVarCommand();
-};
