+++ DefineVarCommand.h
@@ -1,0 +1,11 @@
+#ifndef FLIGHTSIMULATOR__DefineVarCOMMAND_H_
+#define FLIGHTSIMULATOR__DefineVarCOMMAND_H_
+#include Command.h
+class DefineVarCommand : Command {
+	public:
+int execute(string str) {
+
+}
+~DefineVarCommand();
+};
+#endif
