+++ LoopCommand.h
@@ -1,0 +1,10 @@
+#ifndef FLIGHTSIMULATOR__LoopCOMMAND_H_
+#define FLIGHTSIMULATOR__LoopCOMMAND_H_
+#include Command.h
+class LoopCommand : Command {
+int execute(string str) {
+
+}
+~LoopCommand();
+};
+#endif
