+++ OpenServerCommand.h
@@ -1,3 +1,4 @@
+#include OpenServerCommand.h
 class OpenServerCommand : Command {
 public:
 int execute(string str) {
