+++ PrintCommand.h
@@ -1,0 +1,11 @@
+#ifndef FLIGHTSIMULATOR__PrintCOMMAND_H_
+#define FLIGHTSIMULATOR__PrintCOMMAND_H_
+#include Command.h
+class PrintCommand : Command {
+public:
+int execute(string str) {
+
+}
+virtual ~PrintCommand();
+};
+#endif //FLIGHTSIMULATOR__PrintCOMMAND_H_
