+++ SleepCommand.h
@@ -1,0 +1,11 @@
+#ifndef FLIGHTSIMULATOR__SleepCOMMAND_H_
+#define FLIGHTSIMULATOR__SleepCOMMAND_H_
+#include Command.h
+class SleepCommand : Command {
+	public:
+int execute(string str) {
+
+}
+~SleepCommand();
+};
+#endif
