+++ main.cpp
@@ -1,47 +1,6 @@
 // Copyright 2019 Meni Ashurov
-#include <fstream>
-#include <map>
-#include <string>
-#include <iostream>
-#include <list>
-#include <utility>
-#include <fstream>
-#include <sstream>
-#include <vector>
-#include <regex>
-using namespace std;
 
 
-
-class DefineVarCommand : Command {
-int execute(string str) {
-
-}
-~DefineVarCommand();
-};
-class Var_Data {
-
- private:
-  double _value;
-  string _sim;
-
- public:
-  Var_Data(double value, string sim) :
-      _value(value), _sim(sim) {}
-      string get_sim() {
-        return this->_sim;
-      }
-      void set_sim(string sim) {
-        this->_sim = sim;
-      }
-  double get_value() {
-    return this->_value;
-  }
-  void set_value(double value) {
-    this->_value = value;
-  }
-  ~Var_Data(){}
-};
 void check_line(string &line, vector<string> &str_array) {
   const char* token = &(line[0]);
   const char* start = token;
